requirements.txt
discord
aiohttp
RPi.GPIO
numpy